requirements.txt
black
flake8
ruff
mypy
pylint
isort
radon
pytest-xdist